Add risk evaluation and Dockerfile

+++ src/main/java/com/medilabosolutions/front/controllers/UserController.java
@@ -92,7 +92,7 @@ public class UserController {
 	}
 
 	@PostMapping("/update/{id}")
-	public String updateUser(@PathVariable("id") Integer id, @Valid User user, BindingResult result, Model model) {
+	public String updateUser(@PathVariable("id") Long id, @Valid User user, BindingResult result, Model model) {
 		if (result.hasErrors()) {
 			return "update";
 		}
